update for warning

+++ src/state_lattice_planner.cpp
@@ -480,7 +480,7 @@ void StateLatticePlanner::process(void)
                     geometry_msgs::Twist cmd_vel;
                     double calculation_time = ros::Time::now().toSec() - start;
                     int delayed_control_index = std::min(std::ceil(calculation_time * HZ) + CONTROL_DELAY, (double)trajectory.trajectory.size());
-                    if(trajectory.trajectory.size() < CONTROL_DELAY){
+                    if((int)trajectory.trajectory.size() < CONTROL_DELAY){
                         delayed_control_index = std::ceil(calculation_time * HZ);
                     }
                     std::cout << calculation_time << ", " << delayed_control_index << std::endl;
